# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/mode_of_payments:
    get:
      tags:
        - ModeOfPayments
      summary: List ModeOfPayments
      description: Get a paginated list of ModeOfPayments with optional filtering and sorting
      operationId: listModeOfPayments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModeOfPaymentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ModeOfPayments
      summary: Create ModeOfPayment
      description: Create a new ModeOfPayment
      operationId: createModeOfPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModeOfPaymentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModeOfPayment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/payment_allocations:
    get:
      tags:
        - PaymentAllocations
      summary: List PaymentAllocations
      description: Get a paginated list of PaymentAllocations with optional filtering and sorting
      operationId: listPaymentAllocations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentAllocationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PaymentAllocations
      summary: Create PaymentAllocation
      description: Create a new PaymentAllocation
      operationId: createPaymentAllocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentAllocationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentAllocation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ModeOfPayment:
      type: object
      required:
        - id
        - code
        - name
        - mode_type
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 120
        mode_type:
          type: string
        default_account_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    CreateModeOfPaymentRequest:
      type: object
      required:
        - code
        - name
        - mode_type
        - status
        - metadata
      properties:
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 120
        mode_type:
          type: string
        default_account_id:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    ModeOfPaymentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ModeOfPayment'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaymentAllocation:
      type: object
      required:
        - id
        - company_id
        - payment_id
        - invoice_ref
        - invoice_kind
        - allocated_amount
        - discount_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        payment_id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
        allocated_amount:
          type: string
        discount_amount:
          type: string
        discount_account_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreatePaymentAllocationRequest:
      type: object
      required:
        - company_id
        - payment_id
        - invoice_ref
        - invoice_kind
        - allocated_amount
        - discount_amount
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        payment_id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
        allocated_amount:
          type: string
        discount_amount:
          type: string
        discount_account_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    PaymentAllocationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentAllocation'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []